requirements.txt
asgiref==3.4.1
Django==3.2.7
# numpy==1.21.2
# opencv-python==4.5.3.56
pytz==2021.1
sqlparse==0.4.2
mysqlclient==2.1.0